+++ $notes/changelog.md
@@ -22,6 +22,8 @@
   - Crops plantable in botany pots
 - Cosmetic Armors
   - Added the bunny outfit into collection quests and Elna's shop
+- Botania:
+  - world gen flowers are added to flora collection
 ## Mods Removed:
 - Botany Pots Tiers
 - chisels and bits